+++ lib/lightning/collections/collection.ex
@@ -18,6 +18,7 @@ defmodule Lightning.Collections.Collection do
   schema "collections" do
     field :name, :string
     field :byte_size_sum, :integer
+    field :raw_name, :string, virtual: true
     belongs_to :project, Lightning.Projects.Project
     has_many :items, Lightning.Collections.Item
 
@@ -35,5 +36,32 @@ defmodule Lightning.Collections.Collection do
     |> unique_constraint([:name],
       message: "A collection with this name already exists"
     )
+  end
+
+  defp validate_changeset(changeset) do
+    changeset
+    |> validate_format(:name, ~r/^[a-z0-9]+([\-_.][a-z0-9]+)*$/,
+      message: "Collection name must be URL safe"
+    )
+    |> unique_constraint([:name],
+      message: "A collection with this name already exists"
+    )
+  end
+
+  def form_changeset(collection, attrs) do
+    collection
+    |> cast(attrs, [:raw_name])
+    |> validate_required([:raw_name])
+    |> then(fn changeset ->
+      case get_change(changeset, :raw_name) do
+        nil ->
+          changeset
+
+        raw_name ->
+          changeset
+          |> put_change(:name, Lightning.Helpers.url_safe_name(raw_name))
+      end
+    end)
+    |> validate_changeset()
   end
 end